+++ examples/own_commands_ex.py
@@ -11,6 +11,7 @@ from pylatex.base_classes import Environment, CommandBase
 from pylatex.parameters import Arguments
 from pylatex.package import Package
 from pylatex import Document, Section, Command
+from pylatex.utils import NoEscape
 
 
 class ExampleEnvironment(Environment):
@@ -42,15 +43,15 @@ class ExampleCommand(CommandBase):
 # Create a new document
 doc = Document()
 with doc.create(Section('Custom commands')):
-    doc.append(
-        """
-        The following is a demonstration of a custom \\LaTeX
+    doc.append(NoEscape(
+        r"""
+        The following is a demonstration of a custom \LaTeX{}
         command with a couple of parameters.
-        """)
+        """))
 
     # Define the new command
-    new_comm = Command('newcommand{\\exampleCommand}', options=3,
-                       arguments='\\color{#1} #2 #3 \\color{black}')
+    new_comm = Command('newcommand', '\exampleCommand', options=3,
+                       extra_arguments=r'\color{#1} #2 #3 \color{black}')
     doc.append(new_comm)
 
     # Use our newly created command with different arguments
@@ -59,26 +60,26 @@ with doc.create(Section('Custom commands')):
     doc.append(ExampleCommand(arguments=Arguments('red', 'Hello', 'World!')))
 
 with doc.create(Section('Custom environments')):
-    doc.append(
-        """
-        The following is a demonstration of a custom \\LaTeX
+    doc.append(NoEscape(
+        r"""
+        The following is a demonstration of a custom \LaTeX{}
         environment using the mdframed package.
-        """)
+        """))
 
     # Define a style for our box
     mdf_style_definition = Command('mdfdefinestyle',
                                    arguments=['my_style',
-                                              'linecolor=#1,\
-                                               linewidth=#2,\
-                                               leftmargin=1cm,\
-                                               leftmargin=1cm'])
+                                              ('linecolor=#1,'
+                                               'linewidth=#2,'
+                                               'leftmargin=1cm,'
+                                               'leftmargin=1cm')])
 
     # Define the new environment using the style definition above
-    new_env = Command('newenvironment{exampleEnvironment}', options=2,
-                      arguments=[
+    new_env = Command('newenvironment', 'exampleEnvironment', options=2,
+                      extra_arguments=[
                           mdf_style_definition.dumps() +
-                          '\\begin{mdframed}[style=my_style]',
-                          '\\end{mdframed}'])
+                          r'\begin{mdframed}[style=my_style]',
+                          r'\end{mdframed}'])
     doc.append(new_env)
 
     # Usage of the newly created environment