+++ pylatex/__init__.py
@@ -18,3 +18,4 @@ from .figure import Figure, SubFigure
 from .lists import Enumerate, Itemize, Description
 from .quantities import Quantity
 from .base_classes import Command
+from .utils import NoEscape